pyproject.toml
[project]
name = "django-cache-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-redis>=5.4.0",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "faker>=33.3.0",
]
